Split features to separate files
