+++ docker-compose.yml
@@ -8,7 +8,7 @@ services:
         - NODE_ENV=development
     # you can use standard debug config or experimental node inspect
     #command: node --debug=5858 ../node_modules/nodemon/bin/nodemon.js
-    command: node --inspect ../node_modules/nodemon/bin/nodemon.js server.js --exec node
+    command: node --inspect ../node_modules/nodemon/bin/nodemon.js server.js --exec babel-node
     ports:
       - "80:80"
       - "5858:5858"
